agents: Huge refactor to **really** make it work.
This refactors a lot of the logic used to implement the behaviour, to
correctly implement the statement.

The most annoying thing was to give preference to some items of the
queue, so I had to use performatives to filter easily.

It also has a few more tweaks, mainly results are sorted now, and other
nits like that.